Rename baselinetest auth_test to mirror handlers dir structure
